# 05/openapi.yaml
openapi: 3.0.3
info:
  title: AM1 - Homework 5
  version: 1.0.0
...
paths:
...
  /tour:
    post:
      tags:
        - tour
      summary: Add new tour
      description: Add new tour
      operationId: postTour
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tour'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tour'
          links:
            Self:
              $ref: '#/components/links/PostTour'
            GetLocation:
              $ref: '#/components/links/GetTour'
        '400':
          description: Bad request
    get:
      tags:
        - tour
      summary: Get list of all tours
      description: Get list of all tours
      operationId: getTours
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tour'  
          links:
            Self:
              $ref: '#/components/links/GetTours'
            GetTour:
              $ref: '#/components/links/GetTour'
            PostTour:
              $ref: '#/components/links/PostTour'
        '400':
          description: Bad request
  /tour/{id}:
    get:
      tags:
        - tour
      summary: Get tour by id
      description: Get tour by id
      operationId: getTour
      parameters:
        - name: id
          in: path
          description: Tour ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Tour'  
          links:
            Self:
              $ref: '#/components/links/GetTour'
            Next:
              $ref: '#/components/links/GetTour'
            Update:
              $ref: '#/components/links/UpdateTour'
            Delete:
              $ref: '#/components/links/DeleteTour'
        '400':
          description: Bad request
    put:
      tags:
        - tour
      summary: Update tour by id
      description: Update tour by id
      operationId: updateTour
      parameters:
        - name: id
          in: path
          description: Tour ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Tour'  
          links:
            Self:
              $ref: '#/components/links/UpdateTour'
            GetTour:
              $ref: '#/components/links/GetTour'
        '400':
          description: Bad request
    delete:
      tags:
        - tour
      summary: Delete tour by id
      description: Delete tour by id
      operationId: deleteTour
      parameters:
        - name: id
          in: path
          description: Tour ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          links:
            GetTour:
              $ref: '#/components/links/GetTour'
        '400':
          description: Bad request
...
components:
    schemas:
      Country:
        type: object
        properties:
          id:
            type: integer
          name:
            type: string
      Location:
        type: object
        properties:
          id:
            type: integer
          name:
            type: string
          country:
            $ref: '#/components/schemas/Country'
      Tour:
        type: object
        properties:
          id:
            type: integer
          name:
            type: string
          location:
            $ref: '#/components/schemas/Location'
          customers:
            items:
              $ref: '#/components/schemas/Customer'
            type: array
      Customer:
        type: object
        properties:
          id:
            type: integer
          name:
            type: string
    links:
...
      PostTour:
        operationId: postTour
      GetTours:
        operationId: getTours
      GetTour:
        operationId: getTour
        parameters:
          tourId: '$response.body#/id'
      UpdateTour:
        operationId: putTour
        parameters:
          tourId: '$response.body#/id'
      DeleteTour:
        operationId: deleteTour
        parameters:
          tourId: '$response.body#/id'
...
servers:
  - url: 'http://t-a.org'
    variables: {}